// src/containers/AdditionalPOForms/SubcontractAgreement/enum.ts
export enum PO_SUBCONTRACT_AGREEMENT_KEY {
  SUBCONTRACTOR = 'subcontractor',
  DATE = 'date',
  PROJECT = 'project',
  SUBCONTRACTOR_NAME = 'subcontractorName',
  BUSINESS_ADDRESS_AND_TAX_ID_NUMBER = 'businessAddressAndTaxIdNumber',
  CONTRACT_NUMBER = 'contractNumber',
  GRANT_NUMBER = 'grantNumber',
  START_DATE = 'startDate',
  END_DATE = 'endDate',
  EXPENDITURES = 'expenditures',
  EXECUTED_DATE = 'executedDate',
  PRINCIPAL_INVESTIGATOR = 'principalInvestigator',
  SUBCONTRACTOR_SIGNATURE = 'subcontractorSignature',
  RCUH_SIGNATURE = 'rcuhSignature',
  SOC_SUBCONTRACTOR_NAME = 'socSubcontractorName',
  SOC_SUBCONTRACTOR_DATE = 'socSubcontractorDate',
  SOC_SUBCONTRACTOR_SIGNATURE = 'socSubcontractorSignature',
  SOC_SUBCONTRACTOR_TITLE = 'socSubcontractorTitle',
}